+++ Classes/GHCommits.h
@@ -5,5 +5,6 @@
 
 @interface GHCommits : GHCollection
 - (id)initWithRepository:(GHRepository *)repo;
+- (id)initWithRepository:(GHRepository *)repo sha:(NSString *)sha;
 - (id)initWithPullRequest:(GHPullRequest *)pullRequest;
 @end