+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Additions:
 * Set app badge to number of unread notifications
 * Separate icon for forked repositories
 * Added list of repository contributors
+* Added repository commits for branches
 
 Changes: